--- sampleapp/app/home/(tabs)/about.tsx
@@ -1,30 +1,0 @@
-import { View, Text, StyleSheet } from 'react-native';
-
-export default function About() {
-  return (
-    <View style={styles.container}>
-      <Text style={styles.title}>About Us</Text>
-      <Text style={styles.text}>This is the about page of the app.</Text>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    padding: 20,
-    backgroundColor: 'white',
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: 'bold',
-  },
-  text: {
-    fontSize: 16,
-    marginTop: 10,
-  },
-});
-
-
